The SetDatas method now gets the 2 best heroes of each roles and put them into an array

+++ DiscordBot/Entities/Hero.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace DiscordBot.Entities{
     public class Hero{
         public string name;
-        public int timePlayed;
+        public float timePlayed;
         public float winRate;
         public string role;
